[#23] Move AST to package `ast.pure`
Because we want to split the AST in each step, we need to move the base
to a specific package.

This commit do the following:

* Move the AST to package `ast.pure`
* For simplification, remove `ParserUtils` and put methods in `CommonParser`
