+++ src/main/scala/definiti/core/ast/package.scala
@@ -1,0 +1,7 @@
+package definiti.core
+
+package object ast {
+  private[core] type ImportsMap = Map[String, String]
+
+  private[core] val emptyImportsMap: ImportsMap = Map.empty[String, String]
+}
