+++ src/main/scala/definiti/core/parser/project/CommonParser.scala
@@ -1,9 +1,13 @@
 package definiti.core.parser.project
 
-import definiti.core.{ParameterDefinition, TypeReference}
+import definiti.core.ast.pure._
+import definiti.core.ast.{Position, Range}
 import definiti.core.parser.antlr.DefinitiParser.{GenericTypeContext, GenericTypeListContext, ParameterDefinitionContext, ParameterListDefinitionContext}
 import definiti.core.utils.CollectionUtils.scalaSeq
-import definiti.core.utils.ParserUtils.getRangeFromContext
+import org.antlr.v4.runtime.tree.TerminalNode
+import org.antlr.v4.runtime.{ParserRuleContext, Token}
+
+import scala.collection.mutable.ListBuffer
 
 trait CommonParser {
   def processParameterListDefinition(context: ParameterListDefinitionContext): Seq[ParameterDefinition] = {
@@ -31,5 +35,60 @@ trait CommonParser {
       context.IDENTIFIER().getText,
       processGenericTypeList(context.genericTypeList())
     )
+  }
+
+  def extractStringContent(string: String): String = {
+    var temporaryResult = string
+    if (temporaryResult.startsWith("\"")) {
+      temporaryResult = temporaryResult.substring(1)
+    }
+    if (temporaryResult.endsWith("\"")) {
+      temporaryResult = temporaryResult.substring(0, temporaryResult.length - 1)
+    }
+    temporaryResult
+  }
+
+  def extractDocComment(string: String): String = {
+    var temporaryResult = string
+    if (temporaryResult.startsWith("/**")) {
+      temporaryResult = temporaryResult.substring(3)
+    }
+    if (temporaryResult.endsWith("*/")) {
+      temporaryResult = temporaryResult.substring(0, temporaryResult.length - 2)
+    }
+    temporaryResult
+  }
+
+  def getRangeFromContext(context: ParserRuleContext): Range = {
+    def position(token: Token): Position = {
+      if (token == null) {
+        Position(0, 0)
+      } else {
+        Position(token.getLine, token.getCharPositionInLine)
+      }
+    }
+
+    Range(position(context.getStart), position(context.getStop))
+  }
+
+  def getRangeFromTerminalNode(terminalNode: TerminalNode): Range = {
+    val symbol = terminalNode.getSymbol
+    Range(
+      Position(symbol.getLine, symbol.getStartIndex),
+      Position(symbol.getLine, symbol.getStopIndex)
+    )
+  }
+
+  def getRangeFromToken(token: Token): Range = {
+    Range(
+      Position(token.getLine, token.getStartIndex),
+      Position(token.getLine, token.getStopIndex)
+    )
+  }
+
+  def appendIfDefined[A, B](element: A, buffer: ListBuffer[B], transformer: A => B): Unit = {
+    if (element != null) {
+      buffer.append(transformer(element))
+    }
   }
 }
