+++ src/test/scala/definiti/core/generators/ExpressionGenerator.scala
@@ -1,7 +1,8 @@
 package definiti.core.generators
 
 import definiti.core.generators.ASTGenerator.anyRange
-import definiti.core.{BooleanValue, Context, Expression, FunctionCall}
+import definiti.core._
+import definiti.core.ast.pure._
 import org.scalacheck.Gen
 
 object ExpressionGenerator {
