+++ src/test/scala/definiti/core/generators/TypeGenerator.scala
@@ -1,6 +1,7 @@
 package definiti.core.generators
 
 import definiti.core._
+import definiti.core.ast.pure._
 import org.scalacheck.Gen
 
 object TypeGenerator {
