+++ src/test/scala/definiti/core/parser/TestConstants.scala
@@ -1,6 +1,6 @@
 package definiti.core.parser
 
-import definiti.core.{Position, Range}
+import definiti.core.ast.{Position, Range}
 
 object TestConstants {
   val defaultRange: Range = Range(Position(0, 0), Position(0, 0))
