+++ src/test/scala/definiti/core/parser/project/ProcessTypeVerificationSpec.scala
@@ -1,7 +1,8 @@
 package definiti.core.parser.project
 
+import definiti.core.ConfigurationMock
+import definiti.core.ast.pure._
 import definiti.core.generators.antlr.TypeContextGenerator
-import definiti.core.{ConfigurationMock, TypeReference}
 import org.scalacheck.Gen
 import org.scalatest.prop.PropertyChecks
 import org.scalatest.{FlatSpec, Matchers}
